requirements.txt
beautifulsoup4==4.11.1
soupsieve==2.3.2.post1
Pillow==9.2.0
pip==22.2.2
pytz==2022.4
numpy==1.23.3
setuptools==65.4.1
pandas==1.5.0
python-dateutil==2.8.2
certifi==2022.9.24
six==1.16.0
idna==3.4
urllib3==1.26.12
requests==2.28.1